pyproject.toml
[project]
name = "step-by-step"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.0",
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
]
